safeguard-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.9.0
sqlalchemy==2.0.23
sqlite3
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-socketio[asyncio]==5.9.0
websockets==12.0
geopy==2.4.1
haversine==2.8.0
asyncio-mqtt==0.16.1
schedule==1.2.0